python/requirements.txt
bs4
requests
jinja2
pdfkit
python-dotenv